requirements_test.txt
pytest-homeassistant-custom-component>=0.13.218
pytest-cov>=6.0.0
pymodbus>=3.8.3,<4.0.0
